remove dead code and move some things to state/mod.rs
